read me
Make a folder named 'DeRicheImages' on your C drive first. 
You may delete ScreenShot.cs class if you wish. Above codes don't require it. 

all of the following are recommended property changes on FinalForm. None are required for above code to work. 
AutoSize=false; //this reflects what you have in developer mode.(the size of the form)
AutoScrolll=true; //needs to be true for AutoScrollMargin to take effect
AutoScrollMargin = (45,45)  //this creates margins around the panels. not required, recommending it so there would be more room below the adminPanel on the final form.